- All main classes now have new template arguments for the type of indice,    which now defaults to "size_t" instead of "int". In case this breaks    backward compatibility in user code, especify "int" to override the default    template arguments, although "size_t" it's recommended.
- Marked as version 1.1.0 (if no defect is found in a while, a public release will be done).

+++ examples/matrix_example.cpp
@@ -92,7 +92,7 @@ void kdtree_demo(const size_t nSamples,const size_t dim)
 
 	// do a knn search
 	const size_t num_results = 3;
-	vector<int>   ret_indexes(num_results);
+	vector<size_t>   ret_indexes(num_results);
 	vector<num_t> out_dists_sqr(num_results);
 
 	nanoflann::KNNResultSet<num_t> resultSet(num_results);
